HARMONY-1153: Refactor generic parameter parsing code into util/parameter-parsing and out of the the ogc-coverages subset parameter parsing.
