+++ app/frontends/ogc-coverages/get-coverage-rangeset.ts
@@ -4,9 +4,9 @@ import { keysToLowerCase } from '../../util/object';
 import { RequestValidationError } from '../../util/errors';
 import wrap from '../../util/array';
 import { parseVariables } from '../../util/variables';
-import { parsePointParam, parseSubsetParams, subsetParamsToBbox, subsetParamsToTemporal, ParameterParseError } from './util/subset-parameter-parsing';
+import { parsePointParam, parseSubsetParams, subsetParamsToBbox, subsetParamsToTemporal } from './util/subset-parameter-parsing';
 import { parseAcceptHeader } from '../../util/content-negotiation';
-import parseMultiValueParameter from '../../util/parameter-parsing';
+import { ParameterParseError, parseMultiValueParameter } from '../../util/parameter-parsing';
 import HarmonyRequest from '../../models/harmony-request';
 import { createDecrypter, createEncrypter } from '../../util/crypto';
 import parseCRS from '../../util/crs';
